Extracting details to distinct header
